refactored cors configuration logic
